+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Internal/Json/JsonNetSerializer.cs
@@ -88,8 +88,8 @@ namespace Smartsheet.Api.Internal.Json
             // Handles objectValue serialization
             serializer.Converters.Add(new ObjectValueTypeConverter(false));
 
-            // Handles filter value serialization
-            serializer.Converters.Add(new FilterValueTypeConverter());
+            // Handles report filter value serialization
+            serializer.Converters.Add(new ReportFilterValueTypeConverter());
 
             // Handles widget content deserialization
             serializer.Converters.Add(new WidgetContentConverter());